Update README.md
If you just clone and try to run this following the steps, in my experience an error is briefly shown and then it immediately disappears and shows you a TypeScript compiler watch prompt. If you are fast a catch the error before it is cleared, it says "Missing value for AzureWebJobsStorage in local.settings.json. This is required for all triggers other than HTTP. You can run 'func azure functionapp fetch-app-settings <functionAppName>' or specify a connection string in local.settings.json."